requirements.txt
numpy
scipy
opencv-python
pillow
tensorboardX
easydict
matplotlib